+++ CHANGELOG.md
@@ -12,7 +12,10 @@
 - REMOVED: Comparison of several Qwen-1.5 models (removed due to poor scores caused by invalid GGUF)
 - Added benchmark evals for Google Gemini 1.0 Pro (latest version as of 17th Feb 2024)
 - Added benchmark evals for Claude 3 models and MistralAI "mistral-large"
-- Add 3 new test cases to the waitlist: `find_mean`, `find_median` (see `code_generation_waitlist/`).
+- Added benchmark for the latest OpenAI GPT-4 Turbo ("gpt-4-turbo-2024-04-09")
+- Added benchmarks for several open-weights models hosted on Fireworks.ai: Qwen-72b, Mixtral 8x22b (instruct preview), DBRX Instruct
+- Added benchmarks for Llama3 models and Mistral's own Mixtral-8x22b
+- Add 2 new test cases to the waitlist: `find_mean`, `find_median` (see `code_generation_waitlist/`).
 
 ### Fixed
 - Changed the wording from "open-source" models to "locally-hosted" models (a more appropriate description)